Chore #20: Renamed from university to universityselection
