+++ feature/main/src/main/java/com/hankki/feature/main/MainScreen.kt
@@ -42,7 +42,7 @@ import com.hankki.feature.home.navigation.homeNavGraph
 import com.hankki.feature.login.navigation.loginNavgraph
 import com.hankki.feature.my.navigation.myNavGraph
 import com.hankki.feature.report.navigation.reportNavGraph
-import com.hankki.feature.university.navigation.universityNavgraph
+import com.hankki.feature.universityselection.navigation.universitySelectionNavgraph
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toPersistentList
 import kotlinx.coroutines.launch
@@ -87,7 +87,7 @@ internal fun MainScreen(
                         navigateToDummy = { navigator.navigateToDummy() }
                     )
                     loginNavgraph()
-                    universityNavgraph(
+                    universitySelectionNavgraph(
                         navigateToHome = { navigator.navigateToHome() }
                     )
                 }
