+++ feature/universityselection/src/main/AndroidManifest.xml
